% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r02pro.R
\name{r02pro}
\alias{r02pro}
\title{Do the interactive exercises}
\usage{
r02pro(id)
}
\arguments{
\item{id}{the index of the lesson}
}
\value{
This function is an interactive exercise. Hence, no value returned.
}
\description{
This function provides interactive exercises for each lesson corresponding to each subsection of the book "R Programming: Zero to Pro"
}
\examples{
#Do the exercise for Section 1.1
\dontrun{r02pro(1.1)}
}
